webkit/data/layout_tests/platform/chromium-mac/LayoutTests/http/tests/loading/redirect-methods-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
<unknown> - didFinishLoading
frame "0" - didStartProvisionalLoadForFrame
about:blank - willSendRequest <WebRequest URL "about:blank">
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
about:blank - didFinishLoading
frame "0" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-form.html">
frame "0" - didCancelClientRedirectForFrame
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didFinishLoading
frame "0" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php">
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true">
frame "0" - didCommitLoadForFrame
frame "0" - didFinishDocumentLoadForFrame
frame "1" - didStartProvisionalLoadForFrame
about:blank - willSendRequest <WebRequest URL "about:blank">
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
about:blank - didFinishLoading
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - didFinishLoading
frame "1" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-form.html">
frame "1" - didCancelClientRedirectForFrame
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didFinishLoading
frame "1" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php">
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true">
frame "1" - didCommitLoadForFrame
frame "1" - didFinishDocumentLoadForFrame
frame "2" - didStartProvisionalLoadForFrame
about:blank - willSendRequest <WebRequest URL "about:blank">
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
about:blank - didFinishLoading
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - didFinishLoading
frame "2" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-form.html">
frame "2" - didCancelClientRedirectForFrame
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didFinishLoading
frame "2" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php">
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true">
frame "2" - didCommitLoadForFrame
frame "2" - didFinishDocumentLoadForFrame
frame "3" - didStartProvisionalLoadForFrame
about:blank - willSendRequest <WebRequest URL "about:blank">
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
about:blank - didFinishLoading
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - didFinishLoading
frame "3" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-form.html">
frame "3" - didCancelClientRedirectForFrame
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didFinishLoading
frame "3" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php">
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <WebRequest URL "http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true">
frame "3" - didCommitLoadForFrame
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
This test checks to see what HTTP method is used to fetch the final resource in the case where the first request results in a redirect.
301, 302, 303, and 307 http redirects are all tested.
